+++ packages/lumx-react/src/index.ts
@@ -22,6 +22,7 @@ export * from './components/flex-box';
 export * from './components/generic-block';
 export * from './components/heading';
 export * from './components/grid';
+export * from './components/grid-column';
 export * from './components/icon';
 export * from './components/image-block';
 export * from './components/inline-list';