+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.8-alpha-auto-release.0][] - 2022-12-13
+
 ### Added
 
 -   New `GridColumn` component.
@@ -1674,3 +1676,7 @@ _Failed released_
 [3.0.6]: https://github.com/lumapps/design-system/tree/v3.0.6
 [unreleased]: https://github.com/lumapps/design-system/compare/v3.0.7...HEAD
 [3.0.7]: https://github.com/lumapps/design-system/tree/v3.0.7
+
+
+[Unreleased]: https://github.com/lumapps/design-system/compare/v3.0.8-alpha-auto-release.0...HEAD
+[3.0.8-alpha-auto-release.0]: https://github.com/lumapps/design-system/tree/v3.0.8-alpha-auto-release.0